Convert what's new page to be independent of feature
